backend/class-service/requirements.txt
flask==3.0.0
mysql-connector-python==8.2.0
flasgger==0.9.7.1
pika==1.3.2
